fix somebug

+++ Back-End/src/main/java/com/hlteam/naturezen/repository/RoleRepository.java
@@ -1,13 +1,15 @@
 package com.hlteam.naturezen.repository;
 
+import java.util.Optional;
+
 import com.hlteam.naturezen.entity.ERole;
 import com.hlteam.naturezen.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
+
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role,Long> {
     Optional<Role> findByName(ERole name);
 }
